Separate out controls to own component
